[refactor] 프론트엔트 CDN URL 허가 및 postController 전환
