+++ src/main/java/com/finance/finportfolio/domain/post/controller/PostController.java
@@ -2,14 +2,15 @@ package com.finance.finportfolio.domain.post.controller;
 
 import java.util.List;
 
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.finance.finportfolio.domain.post.dto.PostResponseDto;
 import com.finance.finportfolio.domain.post.dto.PostSaveRequestDto;
@@ -19,97 +20,60 @@ import com.finance.finportfolio.domain.post.service.PostService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-// PostController - 게시판 관련 HTTP 요청을 처리
 @Slf4j
-@Controller
-@RequiredArgsConstructor // final 불변성, 컴파일 체크, 테스트 용이
-@RequestMapping("/posts")
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/posts")
 public class PostController {
 
-    // 의존성 주입
     private final PostService postService;
 
-    @GetMapping
-    public String postList(Model model) {
-
-        // 서비스는 Post 엔티티가 아닌 PostResponseDto 리스트를 반환
+    @GetMapping // posts 엔드포인트 접근 시 기본 GET
+    public ResponseEntity<List<PostResponseDto>> getAllPosts() {
         List<PostResponseDto> posts = postService.getAllPosts();
-        model.addAttribute("posts", posts);
-
-        log.info("게시글 목록 조회 완료 - 개수: {}", posts.size());
-        return "posts";
+        log.info("게시글 목록 조회 - 개수: {}", posts.size());
+        return ResponseEntity.ok(posts); // JSON으로 반환
     }
 
-    // post-detail.html return
-    @GetMapping("/detail/{id}")
-    public String toDetailHtml(@PathVariable Long id, Model model) {
-        PostResponseDto dto = postService.getPostById(id);
-
-        log.info("게시글 상세 페이지로 - 제목: {}", dto.getTitle());
-
-        model.addAttribute("post", dto);
-        return "post-detail";
+    @GetMapping("/{id}") // post detail, editor 페이지에서 조회
+    public ResponseEntity<PostResponseDto> getPostById(@PathVariable("id") Long id) {
+        PostResponseDto post = postService.getPostById(id);
+        log.info("게시글 상세 조회 - ID: {}, 제목: {}", id, post.getTitle());
+        return ResponseEntity.ok(post);
     }
 
-    // post-editor.html return (with no ID: new Post)
-    @GetMapping("/editor")
-    public String toEditorHtml(Model model) {
-        model.addAttribute("post", null);
-
-        log.info("게시글 작성 페이지로(with no ID)");
-
-        return "post-editor";
-    }
-
-    @PostMapping("/editor/create")
-    public String createPost(@ModelAttribute PostSaveRequestDto requestDto) {
-
+    @PostMapping // 새 post - portfolio_db에 저장
+    public ResponseEntity<Long> createPost(@RequestBody PostSaveRequestDto requestDto) {
         log.info("게시글 저장 시도 - 제목: {}", requestDto.title());
-
-        // 3. 서비스 호출 시 이미지 전달 (이미 수정한 PostService의 규격에 맞춤)
-        postService.savePost(requestDto);
-
-        return "redirect:/posts";
+        Long postId = postService.savePost(requestDto);
+        return ResponseEntity.status(201).body(postId);
     }
 
-    // post-editor html return (with ID: update)
-    @GetMapping("/editor/{id}")
-    public String toEditorWithID(@PathVariable Long id, Model model) {
-        PostResponseDto dto = postService.getPostById(id); // 기존에 만든 상세조회 활용
-
-        log.info("게시글 수정 페이지로 - 제목: {}", dto.getTitle());
-
-        model.addAttribute("post", dto);
-        return "post-editor";
-    }
-
-    @PostMapping("/editor/{id}")
-    public String updatePost(@PathVariable Long id, @ModelAttribute PostUpdateRequestDto requestDto) {
-
-        log.info("게시글 수정 시도 - 제목: {}", requestDto.title());
-
+    @PatchMapping("/{id}")
+    public ResponseEntity<Long> updatePost(@PathVariable("id") Long id, @RequestBody PostUpdateRequestDto requestDto) {
+        log.info("게시글 수정 시도 - ID: {}, 제목: {}", id, requestDto.title());
         postService.updatePost(id, requestDto);
-        return "redirect:/posts"; // 수정 후 목록으로 리다이렉트
+        return ResponseEntity.ok(id);
     }
 
-    @DeleteMapping("/delete/{id}")
-    public String deletePost(@PathVariable Long id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Long> deletePost(@PathVariable("id") Long id) {
+        log.info("게시글 삭제 시도 - ID: {}", id);
         postService.deletePost(id);
-        log.info("게시글 삭제 완료 - ID: {}", id);
-        return "redirect:/posts";
+        return ResponseEntity.noContent().build();
     }
 
-    // 미참조 이미지 파일 삭제 - 버튼 식(차후 정기 실행으로 변경)
-    @PostMapping("/cleanup")
-    public String cleanUpOrphanFiles() {
+    @DeleteMapping("/cleanup")
+    public ResponseEntity<Void> cleanUpOrphanFiles() {
+        log.info("미참조 이미지 정리 시작");
         postService.cleanUpOrphanFiles();
-
-        return "redirect:/posts";
+        return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/showLog")
-    public String showLog() {
-        log.info("수동 로그");
-        return "redirect:/posts";
+    public ResponseEntity<Void> showLog() {
+        log.info("서버에서 응답!");
+        return ResponseEntity.noContent().build();
     }
+
 }
