--- src/main/java/com/finance/finportfolio/domain/post/controller/PostControllerCSR.java
@@ -1,79 +1,0 @@
-package com.finance.finportfolio.domain.post.controller;
-
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.finance.finportfolio.domain.post.dto.PostResponseDto;
-import com.finance.finportfolio.domain.post.dto.PostSaveRequestDto;
-import com.finance.finportfolio.domain.post.dto.PostUpdateRequestDto;
-import com.finance.finportfolio.domain.post.service.PostService;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/posts")
-public class PostControllerCSR {
-
-    private final PostService postService;
-
-    @GetMapping // posts 엔드포인트 접근 시 기본 GET
-    public ResponseEntity<List<PostResponseDto>> getAllPosts() {
-        List<PostResponseDto> posts = postService.getAllPosts();
-        log.info("게시글 목록 조회 - 개수: {}", posts.size());
-        return ResponseEntity.ok(posts); // JSON으로 반환
-    }
-
-    @GetMapping("/{id}") // post detail, editor 페이지에서 조회
-    public ResponseEntity<PostResponseDto> getPostById(@PathVariable("id") Long id) {
-        PostResponseDto post = postService.getPostById(id);
-        log.info("게시글 상세 조회 - ID: {}, 제목: {}", id, post.getTitle());
-        return ResponseEntity.ok(post);
-    }
-
-    @PostMapping // 새 post - portfolio_db에 저장
-    public ResponseEntity<Long> createPost(@RequestBody PostSaveRequestDto requestDto) {
-        log.info("게시글 저장 시도 - 제목: {}", requestDto.title());
-        Long postId = postService.savePost(requestDto);
-        return ResponseEntity.status(201).body(postId);
-    }
-
-    @PatchMapping("/{id}")
-    public ResponseEntity<Long> updatePost(@PathVariable("id") Long id, @RequestBody PostUpdateRequestDto requestDto) {
-        log.info("게시글 수정 시도 - ID: {}, 제목: {}", id, requestDto.title());
-        postService.updatePost(id, requestDto);
-        return ResponseEntity.ok(id);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Long> deletePost(@PathVariable("id") Long id) {
-        log.info("게시글 삭제 시도 - ID: {}", id);
-        postService.deletePost(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @DeleteMapping("/cleanup")
-    public ResponseEntity<Void> cleanUpOrphanFiles() {
-        log.info("미참조 이미지 정리 시작");
-        postService.cleanUpOrphanFiles();
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/showLog")
-    public ResponseEntity<Void> showLog() {
-        log.info("서버에서 응답!");
-        return ResponseEntity.noContent().build();
-    }
-
-}
