+++ src/main/java/com/finance/finportfolio/global/config/WebConfig.java
@@ -14,7 +14,9 @@ public class WebConfig implements WebMvcConfigurer {
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/api/**") // API 경로에 대해 CORS 적용
-                .allowedOrigins("http://localhost:3000") // React 개발 서버 허용
+                .allowedOrigins("http://localhost:3000",
+                        "https://dh90dx2d0udap.cloudfront.net",
+                        "https://d1wcpvhjz8ef5p.cloudfront.net") // React 개발 서버 허용
                 .allowedMethods("GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS") // 허용할 HTTP 메서드
                 .allowedHeaders("*") // 모든 헤더 허용
                 .allowCredentials(true) // 쿠키/인증 정보 포함 허용
